O3-2812: Appointments: Refactoring and Code Cleanup
