+++ packages/esm-appointments-app/src/home-appointments/appointments-list.component.tsx
@@ -32,7 +32,7 @@ import { EmptyDataIllustration } from './empty-data-illustration.component';
 import { launchCheckInAppointmentModal, handleComplete } from './common';
 import { SeeAllAppointmentsLink, AddAppointmentLink, ViewCalendarLink } from './links.component';
 import { type Appointment, type MappedHomeAppointment } from '../types';
-import { useTodaysAppointments } from './appointments-table.resource';
+import { useTodaysAppointments } from './home-appointments.resource';
 import styles from './appointments-list.scss';
 import { type ConfigObject } from '../config-schema';
